Pulled responders out to its own module.
